move tests that redirect when no session is found into auth code flow test folder move embedded search param check to generic doc request tests
