+++ packages/shopify-app-remix/src/server/authenticate/admin/strategies/__tests__/token-exchange/authenticate.test.ts
@@ -1,4 +1,4 @@
-import {SESSION_COOKIE_NAME, Session} from '@shopify/shopify-api';
+import {Session} from '@shopify/shopify-api';
 
 import {shopifyApp} from '../../../../..';
 import {
